Allow building only HLS project through csynth, skip design export

+++ vivado_hls_build.tcl
@@ -39,21 +39,23 @@ if { [file isdirectory ${PROJ_DIR}/ip/] != 1 } {
 }
 
 # Export the Design
-export_design -flow syn -rtl verilog -format ip_catalog
+if { [info exists ::env(SKIP_EXPORT)] == 0 } {
+   export_design -flow syn -rtl verilog -format ip_catalog
 
 # Copy over the .DCP file
-exec cp -f  [exec ls [glob "${OUT_DIR}/${PROJECT}_project/solution1/impl/verilog/project.runs/synth_1/*.dcp"]] ${PROJ_DIR}/ip/.
+   exec cp -f  [exec ls [glob "${OUT_DIR}/${PROJECT}_project/solution1/impl/verilog/project.runs/synth_1/*.dcp"]] ${PROJ_DIR}/ip/.
 
 # Copy the driver to module source tree
-set DRIVER ${OUT_DIR}/${PROJECT}_project/solution1/impl/ip/drivers
-if { [file exist  ${DRIVER}] } {
-   set DRIVER ${DRIVER}/[exec ls ${DRIVER}]/src
-   set DRIVER [glob ${DRIVER}/*_hw.h]
-   exec cp -f ${DRIVER} ${PROJ_DIR}/ip/.
-}
+   set DRIVER ${OUT_DIR}/${PROJECT}_project/solution1/impl/ip/drivers
+   if { [file exist  ${DRIVER}] } {
+      set DRIVER ${DRIVER}/[exec ls ${DRIVER}]/src
+      set DRIVER [glob ${DRIVER}/*_hw.h]
+      exec cp -f ${DRIVER} ${PROJ_DIR}/ip/.
+   }   
 
 # Copy the HLS report
-exec cp -f  [exec ls [glob "${OUT_DIR}/${PROJECT}_project/solution1/impl/report/verilog/*.rpt"]] ${PROJ_DIR}/ip/.
+   exec cp -f  [exec ls [glob "${OUT_DIR}/${PROJECT}_project/solution1/impl/report/verilog/*.rpt"]] ${PROJ_DIR}/ip/.
+}
 
 # Close current solution
 close_solution
